+++ SleepingStudent/src/components/MemeSelection.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 const MemeSelection = ({ handleRandomize }) => {
   return (
     <div className="meme-selection">
-      <button onClick={handleRandomize}>Generate Random Meme</button>
+      <button className='' onClick={handleRandomize}>Generate Random Meme</button>
     </div>
   );
 };
